+++ src/objects/shared.tsx
@@ -1,0 +1,29 @@
+import {NavigationItem} from '../types';
+
+export const NavigationItems: NavigationItem[] = [
+  {
+    title: 'Balance',
+    path: 'Balance',
+    active: true,
+  },
+  {
+    title: 'Send',
+    path: 'Send',
+    active: false,
+  },
+  {
+    title: 'Receive',
+    path: 'Address',
+    active: false,
+  },
+  {
+    title: 'Status',
+    path: 'Status',
+    active: false,
+  },
+  {
+    title: 'Token',
+    path: 'Token',
+    active: false,
+  },
+];
